moved styles to own files for each control
