+++ src/components/map/FullscreenControl.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import * as L from 'leaflet';
 import { MapProps, Map } from 'react-leaflet';
 import MapStore from '../../stores/mapStore';
-import * as s from './map.scss';
+import * as s from './fullscreenControl.scss';
 import fullScreenEnterIcon from '../../icons/icon-fullscreen-enter.svg';
 import fullScreenExitIcon from '../../icons/icon-fullscreen-exit.svg';
 
@@ -27,7 +27,6 @@ class FullscreenControl extends React.Component<FullscreenControlProps> {
             >
                 <img
                     src={MapStore.isMapFullscreen ? fullScreenExitIcon : fullScreenEnterIcon}
-                    className={s.fullscreenButton}
                 />
             </button>
         );
