+++ src/components/map/coordinateControl.scss
@@ -1,6 +1,36 @@
 @import '../../common.scss';
-.container {
-  position: absolute;
-  right: 10px;
-  top: 10px;
+#coordinateControl {
+  border: 0px;
+  width: 304px;
+  div {
+    height: 29px;
+    width: 130px;
+    display: inline-block;
+    border: 1px solid black;
+    background-color: $lightGrey;
+  }
+  div:first-of-type{
+    margin-right: 40px;
+  }
+  div>button {
+    font-family: $hslFont;
+    color: $fontColorDefault;
+    background: transparent;
+    font-size: $smallFontSize;
+    width: 27px;
+    border: 0px;
+    padding: 0px;
+    margin: 0px;
+  }
+  div>input {
+    font-family: $hslFont;
+    color: $fontColorLightGrey;
+    font-size: $smallFontSize;
+    height: 100%;
+    width: 100px;
+    margin: 0px;
+    border: 0px;
+    border-left: 1px solid black;
+    padding: 0px 0px 0px 2px;
+  }
 }
