+++ src/components/map/measurementControl.scss
@@ -1,0 +1,82 @@
+@import '../../common.scss';
+#measurementControl {
+  height: 30px;
+  background-color: $white;
+  -webkit-box-sizing: content-box;
+  -moz-box-sizing: content-box;
+  box-sizing: content-box;
+
+  div{
+    border:0px;
+    cursor: pointer;
+    height:100%;
+    transition: 0.2s;
+    &:hover {
+      background-color: $lightGrey;
+    }
+    &:first-child {
+      border-top-left-radius: 2px;
+      border-bottom-left-radius: 2px;
+    }
+    &:nth-child(2) {
+      border-left: 1px solid $borderOpaque;
+    }
+    &:nth-child(3) {
+      border-left: 1px solid $borderOpaque;
+      border-top-right-radius: 2px;
+      border-bottom-right-radius: 2px;
+    }
+    svg {
+      stroke: $blue;
+      fill: $blue;
+      height: 30px;
+    }
+  }
+  div.measurementTool {
+    display: inline-block;
+  }
+  div.removeTool {
+    display: none;
+    svg:hover {
+      color: $darkGrey;
+    }
+  }
+  div.removeAll {
+    display: none;
+    svg:hover {
+      color: $darkGrey;
+    }
+  }
+  div.show {
+    display: inline-block;
+  }
+  div.selected {
+    background-color: $blue;
+    svg {
+      stroke: $white;
+      fill: $white;
+    }
+  }
+}
+.measurementCursor {
+  cursor: crosshair;
+}
+
+.movingPolyline {
+  @extend .no-events;
+  stroke: black;
+  stroke-dasharray: 10px;
+}
+
+.polyline {
+  stroke: $darkGrey;
+}
+
+.circleMarker {
+  stroke: black;
+  fill: $whiteOpaque;
+}
+
+.no-events{
+  pointer-events: none;
+}
